Floyd-Warshall algorithm with successors.
compute all-pairs shortest paths with an odd number of edges, along with those with an even number of edges, for a directed graph. These results are computed at the same time and are stored in matrices. The successor matrices must also be maintained.
Zero-edge, i.e. reflexive, paths are not to be considered. Entries in the adjacency matrices should be initialized with a distance of “infinity”. It is acceptable for self-loops to appear in the input file. Input edges are only recorded in the “odd” matrix, i.e. the “even” matrix is initialized for no even paths existing.
3. A vertex may appear several times on a path so that “oddness” or “evenness” may be achieved. This also applies to the starting and destination vertices.